auth and layout changes
