+++ client/src/pages/Layout.tsx
@@ -1,9 +1,11 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
+import Navbar from "../components/Navbar/Navbar";
 
 const Layout = () => {
   return (
     <main>
+      <Navbar />
       <Outlet />
     </main>
   );
